C-ize sources in 'lib/'. Minor interface updates.
* lib/src/div.sl,
* lib/src/io.sl,
* lib/src/roman.sl: Rename to...

* lib/src/div.c,
* lib/src/io.c,
* lib/src/roman.c: ... here. Use CPP syntax where appropriate.

* lib/mtalpha.mk,
* lib/seqc.mk,
* lib/ptl.mk: Update accordingly.

* lib/include/cstdint.h: Do not try to include `<cstdint>' in C++
mode since it is not (yet) defined in ISO C++. Be optimistic and
assume `<stdint.h>' is always available.

* lib/include/svp/io.h,
* lib/include/svp/iomacros.h,
* lib/src/io.c: Use `long' instead of `long long', since `long' is
the canonical largest integer type that fits in a register and we
don't support multi-register types yet.

* tools/include/svp/testoutput.h: Use `likely' and `unlikely'
where appropriate.

+++ lib/include/svp/io.h
@@ -40,7 +40,7 @@ sl_decl(svp_io_putf, void,
 
 /* Print an unsigned integer to the console output. */
 sl_decl(svp_io_putun, void,
-	sl_glparm(unsigned long long, gn),
+	sl_glparm(unsigned long, gn),
 	sl_glparm(unsigned, gbase));
 
 /* Print a nul-terminated string to the console output. */
@@ -56,7 +56,7 @@ sl_decl(svp_io_putf, void,
 
 /* Print a signed integer to the console output. */
 sl_decl(svp_io_putn, void,
-	sl_glparm(long long, gn),
+	sl_glparm(long, gn),
 	sl_glparm(unsigned, gbase));
 
 /* Format and print arguments to the console output. */
